ViewModelの機能を移した
